requirements.txt
streamlit>=1.20
playwright
chromadb
openai>=0.27.0
python-dotenv
pydantic
sqlalchemy
langchain
langchain-openai
python-dotenv
langchain-community
langchain-chroma